+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-* 1.5 (2018-09-03)
-    * Use subcommands instead of flags in auniter.sh to simplify the
-      common interactive use cases.
+* 1.6 (2018-09-11)
+    * Support 'monitor' subcommand using an external serial port terminal
+      (e.g. picocom).
+    * Add 'upmon' subcommand, a combination of 'upload' and 'monitor'.
+    * Add '[auniter] baud' parameter to control default baud rate of port.
 * 1.4.1 (2018-09-03)
     * Fix bug which disabled --locking by default.
     * Allow serial port specifier in --boards flag to omit "/dev/tty" prefix.